🎉 update algolia index when publishing/updating/unpublishing gdoc posts
